added join policy, test passing
Need to think a bit about this one. I am using the lower bound
of size_hint. Maybe we want to restrict this to exact sized Iterators?
This will make the base types a bit more complicated...
As a side note, maybe we want to implement size_hint() for Slice Iterator.