Added bench for composed, fixed bugs in Logs and improved Composed
This commit changed how `Composed` behaves: now, instead of waiting
for 90% of the work to be done to allow parallelism, only the
*last* task will be allowed to have inner parallelism.

The `Log` adaptor now has `#[cfg(feature = "logs")]` guards, in order
to prevent its use when the logging isn't available.

A bench for `Composed` was added: it's the same computation as the
points example.