Formatting fixes, experiments done to find performance issue with 2 threads
We found out that with 2 threads, the calculation of the even sum problem
takes much more longer for our Composed adaptor than the sequential inner
version. The even sum problem here had a large number of vectors, who were
very small (100 elements each). We found out that it was partially due to
this that we have the performance issue: it takes roughly 50 ns to sum
100 elements; we have around 10% of performance difference, which means
that we only need 5 ns of difference between the two calculations to induce
the problem.

By experimenting, we found out that in order to mitigate the problem (but
not completely), we can increase the size of the inner vectors to 2000;
there are still performance issues, but this is a start in the good direction.

This commit also fixes some formatting issues from earlier commits where rustfmt
wasn't set up properly (also, CRLF problems from my Windows environment).