Added `collect()` to parallel iterators and fixed a bug with `Range<u64>` divisions
The `collect()` is implemented following [Fred's notes](http://wagnerf.pages.ensimag.fr/algoa/cm5/resume.html);
it's only available for `Vec<T>`, but it's enough for now.

The bug fixed for `Range<u64>` division is related to how we divide ranges with an
odd length. For our algorithms to work efficiently, we need to put the largest half
on the left side; but the previous implementation of the division used an index
calculated with `(end - start) / 2`. The integer division here will take the
floor of the float result, which means that in the case of an odd length, the right
side of the division will be larger by one.

Instead, we now calculate the index using `(end - start) - ((end - start) / 2)`:
we still use integers, and using this we put the largest half of the range on the
left.