try_reduce (basic)
ok, try_reduce is now here (easy version for now).

I had to change things a little bit because the
fold operation when doing a try_reduce is different from the fold op
when doing a reduce.

because of flatmap this is a problem since the inner scheduler
of the flatmap needs to fold with the correct fold op which
means the information about how to fold must flow inside the consumers.

easy solution: add a fold method in the Reducer trait.

the good consequence of that is that the special scheduler for
try_reduce (join) is not needed anymore since all aborting stuff is now
done directly in the Reducer itself.