scheduler trait
ok, there is now a specific trait for schedulers instead
of having just a function.

this way we can store additional data in the scheduler itself.

this is a requirement for a by_blocks scheduler because it needs to
store the inner scheduler used on each block.
we could also use it to store block sizes or any meaningful data.

it also opens the way to communications between schedulers by
adding additional methods in the trait.
we could for example add a method for asking the scheduler above you
what he thinks about you going parallel.