scheduler method: lifetime added
we now have a ByBlocks adaptor.
we could even easily add a size iterator inside, yeah !

I added it into try_reduce but i'm not sure it is a good idea for
experiments.
it might be better to control everything manually for now.

only problem is that the macro blocks divisions count :-( for rayon
policy.

there might be several solutions here but i'm not sure which one is the
good one.

I think it makes sense to call by_blocks just before the reduction.
which producer should it really divide ?
would it make sense to divide the leftmost producer only ?
would it be always what we want ? what about even_levels ?

too tired to continue. was a big day today.