last change to by_blocks: adaptor instead of sched
ok, so i thought during the night.

so by_blocks is now an adaptor only.

if you do for example

range.par_iter().filter().by_blocks().rayon().try_reduce()

then the producer that is cut into blocks is the FilterProducer.
we then have a sequence of FilterProducer and on each of them
we call rayon().try_reduce().

this way rayon works with the blocks and by choosing where you put
the by_blocks call you can control everything very finely.

we can do any kind of experiment we want.

there is a drawback though but I guess it is not important for now:
we cannot automatically add a .by_blocks() into the try_reduce
because flatmap().try_reduce() would turn out as
flatmap().by_blocks().try_reduce() and the blocks would be inside the
flatmap instead of outside.