work changed ; completed removed
so, here are the changes:

1) completed is no more
2) Worker has no more divide and should be divided
instead we take them from S and require S to be Divisible.
work fn is now a method on S.
3) all Workers loop on something now.
4) FoldProducer is now much like WorkerProducer
it contains a state which is the current `init`.
partial_fold also ignores the fold op.

it's untested but it means that:
writing v.par_iter().filter().fold().adaptive().map().reduce()

will do exactly what we expect

what could still be done ?

a) running clippy would be nice because i spotted and removed some useless closures
b) since fold_op is not used in partial_fold it is most likely not the
right api ?
c) completed closure in work is not nice. what we really need is a size
closure: Fn(&Self) -> (usize, Option<usize>). this way we could work by blocks and not lie in sizes we give.
because currently WorkerProducer::sizes is not telling the truth