Kill the AdaptiveProducer trait and Blocked; all are adaptive now (#1)
* Kill the AdaptiveProducer trait and Blocked; all are adaptive now
There are some things that might be made faster (I am pretty sure they
didnt get slower)
Mutable slices: It is the real iterator, can't divide that thing with
just a mutable reference
Fold: Should probably divide_at, then fold sequentially the left side in
one go, then do a "partial_fold" on the output. However, what if
divide_at on the base doesn't work? (there are some cases which panic)

Some things I am not sure about:
What about the filter? limit for partial fold is on the real elements or
the filtered elements?

* work changed ; completed removed

so, here are the changes:

1) completed is no more
2) Worker has no more divide and should be divided
instead we take them from S and require S to be Divisible.
work fn is now a method on S.
3) all Workers loop on something now.
4) FoldProducer is now much like WorkerProducer
it contains a state which is the current `init`.
partial_fold also ignores the fold op.

it's untested but it means that:
writing v.par_iter().filter().fold().adaptive().map().reduce()

will do exactly what we expect

what could still be done ?

a) running clippy would be nice because i spotted and removed some useless closures
b) since fold_op is not used in partial_fold it is most likely not the
right api ?
c) completed closure in work is not nice. what we really need is a size
closure: Fn(&Self) -> (usize, Option<usize>). this way we could work by blocks and not lie in sizes we give.
because currently WorkerProducer::sizes is not telling the truth

* removed lints and warnings

all but one.

flat_map is dead though. added a test for it

* fix for block sizes

Co-authored-by: frederic wagner <frederic.wagner@imag.fr>