rev + switched slice iter to standard one
ok, here are a few changes.

first we needed to implement rev in order to have reversed parallel
iterators.

nothing difficult but this requires Producers to implement
DoubleEndedIterator hence the large number of files modified.

note that this will require most types to implement rfold.
we also need a partial_rfold.
since anyway we did not even implement fold for everyone,
it is still on the todo list.

-----

second thing: in order to ease my life i switched the slice iterators
from our custom version to rust's standard implementation.
maybe this will impact the sort's performances.

however if this really is the case doesn't it means that we could
improve the standard slice iterator ?
it would be kind of a big deal

----

i re-implemented partial_fold on the slices IterMut
(should be better performances)

I found a nice crate which does exactly what i want and i think
we could reuse it in other places in order
to avoid both Options and unsafe.